requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
cryptography==41.0.4
pytest==7.4.2
pytest-flask==1.3.0
colorama==0.4.6
termcolor==2.3.0 